+++ src/muscle_mem/engine.py
@@ -1,5 +1,4 @@
 import ast
-import time
 import functools
 import hashlib
 import inspect
@@ -10,11 +9,13 @@ from typing import Any, Callable, Dict, List, Optional, ParamSpec, Tuple, TypeVa
 from colorama import Fore, Style
 
 from .check import Check
+from .metrics import Metrics
 from .persistence import DB
 from .types import Step, Trajectory
 
 P = ParamSpec("P")
 R = TypeVar("R")
+
 
 def hash_ast(func):
     # Hashes the ast of a function, to raise errors if implementation changes from persisted tools
@@ -81,97 +82,6 @@ class Tools:
 
         return tool
 
-class Metrics:
-    def __init__(self):
-        self.enabled = False
-        self.metrics = {
-            "query": {"total_time": 0, "count": 0},
-            "filter": {
-                "partials": {"total_time": 0, "count": 0},
-                "capture": {"total_time": 0, "count": 0},
-                "compare": {"total_time": 0, "count": 0}
-            },
-            "runtime": {
-                "precheck": {
-                    "capture": {"total_time": 0, "count": 0},
-                    "compare": {"total_time": 0, "count": 0}
-                },
-                "postcheck": {
-                    "capture": {"total_time": 0, "count": 0},
-                    "compare": {"total_time": 0, "count": 0}
-                }
-            }
-        }
-    
-    def enable(self):
-        self.enabled = True
-    
-    def disable(self):
-        self.enabled = False
-    
-    def report(self):
-        """Print a pretty-printed JSON report of all metrics."""
-        if not self.enabled:
-            return
-        
-        # Create a copy of metrics with averages and formatted times added
-        import copy
-        import json
-        
-        report_metrics = copy.deepcopy(self.metrics)
-        
-        # Format time in appropriate units and add averages
-        def format_time(seconds):
-            """Format time in appropriate units based on magnitude."""
-            if seconds >= 1.0:
-                return f"{seconds:.4f}s"  # seconds
-            elif seconds >= 0.001:
-                return f"{seconds * 1000:.4f}ms"  # milliseconds
-            else:
-                return f"{seconds * 1000000:.4f}us"  # microseconds
-        
-        def process_metrics(metrics_dict):
-            for key, value in metrics_dict.items():
-                if isinstance(value, dict):
-                    if "total_time" in value and "count" in value:
-                        # Store original time values for calculation
-                        raw_time = value["total_time"]
-                        
-                        # Format the time value
-                        value["total_time"] = format_time(raw_time)
-                        
-                        # Calculate and format average if count > 0
-                        if value["count"] > 0:
-                            avg_time = raw_time / value["count"]
-                            value["avg_time"] = format_time(avg_time)
-                    else:
-                        process_metrics(value)
-        
-        process_metrics(report_metrics)
-        print(json.dumps(report_metrics, indent=4))
-
-    def reset(self):
-        """Reset all metrics to zero."""
-        self.__init__()
-    
-    @contextmanager
-    def measure(self, *path):
-        if not self.enabled:
-            yield
-            return
-        
-        start = time.time()
-        try:
-            yield
-        finally:
-            # Navigate to the correct spot in the metrics dictionary
-            current = self.metrics
-            for part in path[:-1]:
-                current = current[part]
-            
-            # Update time and count
-            current[path[-1]]["total_time"] += time.time() - start
-            current[path[-1]]["count"] += 1
 
 class Engine:
     def __init__(self):
@@ -281,7 +191,6 @@ class Engine:
     def filter_pre_checks(self, candidates: List[Trajectory], idx: int) -> List[Trajectory]:
         """Filter trajectories to only those where the next step passes pre-checks"""
 
-
         selected = []
         for candidate in candidates:
             if idx >= len(candidate.steps):
@@ -321,7 +230,6 @@ class Engine:
 
         step_idx = 0
         while True:
-
             # Attempt to top up trajectories if we've run out
             if not trajectories:
                 with self.metrics.measure("query"):
